requirements.txt
alembic==1.0.6
amqp==2.4.0
bcrypt==3.1.6
billiard==3.5.0.5
celery==4.2.1
cffi==1.11.5
Click==7.0
elasticsearch==6.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
Mako==1.0.7
MarkupSafe==1.1.0
Pillow==5.4.1
pycparser==2.19
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.9
redis==3.0.1
six==1.12.0
SQLAlchemy==1.2.16
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
WTForms==2.2.1
